WA-449 fix failing ui tests for PendingSafeViewController
